* Typecasted time_t references to their appropriate types (%ld being   signed long, %lu being unsigned long).  Yes, there are some %lds,   believe it or not.
git-svn-id: http://svn.ratbox.org/svnroot/trunk@7080 b93f080a-4cfa-0310-a19a-80a68f531ef9

+++ modules/core/m_sjoin.c
@@ -254,8 +254,12 @@ static void ms_sjoin(struct Client *client_p,
 
   if (newts < 800000000)
     {
-      sendto_realops_flags(FLAGS_DEBUG, "*** Bogus TS %lu on %s ignored from %s",
-			   newts, chptr->chname, client_p->name);
+      sendto_realops_flags(FLAGS_DEBUG,
+			"*** Bogus TS %lu on %s ignored from %s",
+			(unsigned long) newts,
+			chptr->chname,
+			client_p->name);
+
       newts = oldts;
     }
 
@@ -369,8 +373,10 @@ static void ms_sjoin(struct Client *client_p,
       modebuf[1] = '\0';
     }
 
-  ircsprintf(buf, ":%s SJOIN %lu %s %s %s :", parv[0], tstosend, parv[2],
-          modebuf, parabuf);
+  ircsprintf(buf, ":%s SJOIN %lu %s %s %s :",
+	parv[0],
+	(unsigned long) tstosend,
+	parv[2], modebuf, parabuf);
 
   mbuf = modebuf;
   para[0] = para[1] = para[2] = para[3] = "";
