+++ modules/m_lljoin.c
@@ -250,20 +250,29 @@ static void ms_lljoin(struct Client *client_p,
         }
 
       sendto_one(uplink,
-		 ":%s SJOIN %lu %s + :@%s", me.name,
-		 chptr->channelts, chptr->chname, nick);
+		 ":%s SJOIN %lu %s + :@%s",
+		 me.name,
+		 (unsigned long) chptr->channelts,
+		 chptr->chname,
+		 nick);
     }
   else if ((flags == CHFL_HALFOP) && (IsCapable(uplink, CAP_HOPS)))
     {
       sendto_one(uplink,
-		 ":%s SJOIN %lu %s + :%%%s", me.name,
-		 chptr->channelts, chptr->chname, nick);      
+		 ":%s SJOIN %lu %s + :%%%s",
+		 me.name,
+		 (unsigned long) chptr->channelts,
+		 chptr->chname,
+		 nick);
     }
   else
     {
       sendto_one(uplink,
-		 ":%s SJOIN %lu %s + :%s", me.name,
-		 chptr->channelts, chptr->chname, nick);
+		 ":%s SJOIN %lu %s + :%s",
+		 me.name,
+		 (unsigned long) chptr->channelts,
+		 chptr->chname,
+		 nick);
     }
 
   add_user_to_channel(chptr, target_p, flags);
